refactor : usePostMakeTeamHandler useEffect 제거했습니다
